requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
geopy==2.4.1
matplotlib==3.8.2
pandas==2.1.4
numpy==1.26.2
seaborn==0.13.0
python-dotenv==1.0.0
gunicorn